requirements.txt
python-telegram-bot==22.1
requests==2.32.3
mysql-connector-python==9.3.0
python-dotenv==1.1.0